requirements.txt
streamlit
pandas
plotly
numpy
matplotlib
opencv-python-headless
PyYAML
watchdog
tabulate
requests
google-api-python-client
google-auth